add validation to AccountController

+++ controllers/src/main/java/com/netcracker/komarov/controllers/controller/AccountController.java
@@ -2,7 +2,9 @@ package com.netcracker.komarov.controllers.controller;
 
 import com.google.gson.Gson;
 import com.netcracker.komarov.services.dto.entity.AccountDTO;
+import com.netcracker.komarov.services.dto.entity.ClientDTO;
 import com.netcracker.komarov.services.interfaces.AccountService;
+import com.netcracker.komarov.services.interfaces.ClientService;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -16,23 +18,32 @@ import java.util.Collection;
 public class AccountController {
     private AccountService accountService;
 
+    private ClientService clientService;
 
     @Autowired
-    public AccountController(AccountService accountService) {
+    public AccountController(AccountService accountService,
+                             ClientService clientService) {
         this.accountService = accountService;
+        this.clientService = clientService;
     }
 
     @ApiOperation(value = "Creation of new account")
     @RequestMapping(value = "/clients/{clientId}/accounts", method = RequestMethod.POST)
     public ResponseEntity create(@PathVariable long clientId) {
         Gson gson = new Gson();
-        AccountDTO dto = accountService.createAccount(new AccountDTO(false, 0), clientId);
         ResponseEntity responseEntity;
-        if (dto == null) {
-            responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(gson.toJson("Server error"));
+        ClientDTO clientDTO = clientService.findById(clientId);
+        if (clientDTO == null) {
+            responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(gson.toJson("No such client in database"));
         } else {
-            responseEntity = ResponseEntity.status(HttpStatus.CREATED).body(gson.toJson(dto));
+            AccountDTO dto = accountService.createAccount(new AccountDTO(false, 0), clientId);
+            if (dto == null) {
+                responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                        .body(gson.toJson("Server error"));
+            } else {
+                responseEntity = ResponseEntity.status(HttpStatus.CREATED).body(gson.toJson(dto));
+            }
         }
         return responseEntity;
     }
@@ -41,14 +52,20 @@ public class AccountController {
     @RequestMapping(value = "/admins/requests/accounts/{accountId}", method = RequestMethod.PATCH)
     public ResponseEntity unlock(@PathVariable long accountId) {
         Gson gson = new Gson();
-        AccountDTO dto = accountService.unlockAccount(accountId);
+        AccountDTO accountDTO = accountService.findById(accountId);
         ResponseEntity responseEntity;
-        if (dto == null) {
-            responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(gson.toJson("Server error"));
+        if (accountDTO == null) {
+            responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(gson.toJson("No such admin in database"));
         } else {
-            responseEntity = ResponseEntity.status(HttpStatus.OK)
-                    .body(gson.toJson(dto));
+            AccountDTO dto = accountService.unlockAccount(accountId);
+            if (dto == null) {
+                responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                        .body(gson.toJson("Server error"));
+            } else {
+                responseEntity = ResponseEntity.status(HttpStatus.OK)
+                        .body(gson.toJson(dto));
+            }
         }
         return responseEntity;
     }
@@ -57,14 +74,20 @@ public class AccountController {
     @RequestMapping(value = "/clients/{clientId}/accounts/{accountId}", method = RequestMethod.PATCH)
     public ResponseEntity lock(@PathVariable long clientId, @PathVariable long accountId) {
         Gson gson = new Gson();
-        AccountDTO dto = accountService.lockAccount(accountId);
         ResponseEntity responseEntity;
-        if (dto == null) {
-            responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(gson.toJson("Server error"));
+        ClientDTO clientDTO = clientService.findById(clientId);
+        if (clientDTO == null) {
+            responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(gson.toJson("No such client in database"));
         } else {
-            responseEntity = ResponseEntity.status(HttpStatus.OK)
-                    .body(gson.toJson(dto));
+            AccountDTO dto = accountService.lockAccount(accountId);
+            if (dto == null) {
+                responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                        .body(gson.toJson("Server error"));
+            } else {
+                responseEntity = ResponseEntity.status(HttpStatus.OK)
+                        .body(gson.toJson(dto));
+            }
         }
         return responseEntity;
     }
@@ -73,14 +96,20 @@ public class AccountController {
     @RequestMapping(value = "/clients/{clientId}/accounts/{accountId}/money", method = RequestMethod.PATCH)
     public ResponseEntity refill(@PathVariable long clientId, @PathVariable long accountId) {
         Gson gson = new Gson();
-        AccountDTO dto = accountService.refill(accountId);
         ResponseEntity responseEntity;
-        if (dto == null) {
-            responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(gson.toJson("Server error"));
+        ClientDTO clientDTO = clientService.findById(clientId);
+        if (clientDTO == null) {
+            responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(gson.toJson("No such client in database"));
         } else {
-            responseEntity = ResponseEntity.status(HttpStatus.OK)
-                    .body(gson.toJson(dto));
+            AccountDTO dto = accountService.refill(accountId);
+            if (dto == null) {
+                responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                        .body(gson.toJson("Server error"));
+            } else {
+                responseEntity = ResponseEntity.status(HttpStatus.OK)
+                        .body(gson.toJson(dto));
+            }
         }
         return responseEntity;
     }
@@ -90,14 +119,20 @@ public class AccountController {
     public ResponseEntity getByClientIdAndLock(@PathVariable long clientId, @RequestParam(name = "lock",
             required = false, defaultValue = "false") boolean lock) {
         Gson gson = new Gson();
-        Collection<AccountDTO> dtos = accountService.getAccountsByClientIdAndLock(clientId, lock);
         ResponseEntity responseEntity;
-        if (dtos == null) {
-            responseEntity = ResponseEntity
-                    .status(HttpStatus.INTERNAL_SERVER_ERROR).body(gson.toJson("Server error"));
+        ClientDTO clientDTO = clientService.findById(clientId);
+        if (clientDTO == null) {
+            responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(gson.toJson("No such client in database"));
         } else {
-            responseEntity = ResponseEntity.status(HttpStatus.OK)
-                    .body(gson.toJson(dtos.isEmpty() ? "Empty list of accounts" : dtos));
+            Collection<AccountDTO> dtos = accountService.getAccountsByClientIdAndLock(clientId, lock);
+            if (dtos == null) {
+                responseEntity = ResponseEntity
+                        .status(HttpStatus.INTERNAL_SERVER_ERROR).body(gson.toJson("Server error"));
+            } else {
+                responseEntity = ResponseEntity.status(HttpStatus.OK)
+                        .body(gson.toJson(dtos.isEmpty() ? "Empty list of accounts" : dtos));
+            }
         }
         return responseEntity;
     }
@@ -112,10 +147,47 @@ public class AccountController {
     }
 
     @ApiOperation(value = "Deleting account by ID")
-    @RequestMapping(value = "/accounts/{accountId}", method = RequestMethod.DELETE)
-    public ResponseEntity deleteById(@PathVariable long accountId) {
-        accountService.deleteById(accountId);
+    @RequestMapping(value = "/clients/{clientId}/accounts/{accountId}", method = RequestMethod.DELETE)
+    public ResponseEntity deleteById(@PathVariable long clientId, @PathVariable long accountId) {
         Gson gson = new Gson();
-        return ResponseEntity.status(HttpStatus.OK).body(gson.toJson("Account was deleted"));
+        ResponseEntity responseEntity;
+        ClientDTO clientDTO = clientService.findById(clientId);
+        if (clientDTO == null) {
+            responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(gson.toJson("No such client"));
+        } else {
+            AccountDTO accountDTO = accountService.findById(accountId);
+            if (accountDTO == null) {
+                responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body(gson.toJson("No such account in database"));
+            } else {
+                accountService.deleteById(accountId);
+                responseEntity = ResponseEntity.status(HttpStatus.OK)
+                        .body(gson.toJson("Account was deleted"));
+            }
+        }
+        return responseEntity;
+    }
+
+    @ApiOperation(value = "Selecting account by ID")
+    @RequestMapping(value = "/clients/{clientId}/accounts/{accountId}", method = RequestMethod.GET)
+    public ResponseEntity findById(@PathVariable long clientId, @PathVariable long accountId) {
+        Gson gson = new Gson();
+        ResponseEntity responseEntity;
+        ClientDTO clientDTO = clientService.findById(clientId);
+        if (clientDTO == null) {
+            responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(gson.toJson("No such client in database"));
+        } else {
+            AccountDTO dto = accountService.findById(accountId);
+            if (dto == null) {
+                responseEntity = ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body(gson.toJson("No such account in database"));
+            } else {
+                responseEntity = ResponseEntity.status(HttpStatus.OK)
+                        .body(gson.toJson(dto));
+            }
+        }
+        return responseEntity;
     }
 }
